requirements.txt
simpy
networkx
geopy
flask
flask-limiter
openai
boto3
google-generativeai
python-dotenv
